# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AJN Luxury Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # GALLERY
  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery items
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Gallery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"
    post:
      operationId: createGalleryItem
      tags: [gallery]
      summary: Create gallery item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
  /gallery/{id}:
    delete:
      operationId: deleteGalleryItem
      tags: [gallery]
      summary: Delete gallery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # REVIEWS
...
  /accounting/transactions:
    get:
      operationId: listTransactions
      tags: [accounting]
      summary: List accounting transactions
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: account
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [accounting]
      summary: Create transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer with orders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  # EMPLOYEES
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, fullName, username, role, isActive, permissions]
      properties:
        id:
          type: integer
        name:
          type: string
        fullName:
          type: string
        username:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [owner, admin, staff, delivery]
        isActive:
          type: boolean
        permissions:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Booking:
      type: object
      required: [id, title, customerName, customerPhone, date, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        serviceType:
          type: string
        date:
          type: string
        time:
          type: ["string", "null"]
        status:
          type: string
          enum: [confirmed, awaiting_payment, completed, cancelled]
        color:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    BookingInput:
      type: object
      required: [title, customerName, customerPhone, date]
      properties:
        title:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        serviceType:
          type: string
        date:
          type: string
        time:
          type: string
        notes:
          type: string
    BookingUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
        date:
          type: string
        time:
          type: string
    GalleryItem:
      type: object
      required: [id, type, url, category, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [image, video]
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        category:
          type: string
        titleAr:
          type: ["string", "null"]
        createdAt:
          type: string
    GalleryItemInput:
      type: object
      required: [type, url, category]
      properties:
        type:
          type: string
        url:
          type: string
        thumbnailUrl:
          type: string
        category:
          type: string
        titleAr:
          type: string
...
    Transaction:
      type: object
      required: [id, type, account, amount, description, date, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [receipt, expense, transfer]
        account:
          type: string
          enum: [cash, bank, zain_cash, master_card, asia_hawala]
        toAccount:
          type: ["string", "null"]
        amount:
          type: number
        description:
          type: string
        referenceId:
          type: ["string", "null"]
        date:
          type: string
        createdAt:
          type: string
    TransactionInput:
      type: object
      required: [type, account, amount, description, date]
      properties:
        type:
          type: string
        account:
          type: string
        toAccount:
          type: string
        amount:
          type: number
        description:
          type: string
        referenceId:
          type: string
        date:
          type: string
...
    Customer:
      type: object
      required: [id, name, phone, totalOrders, totalSpent, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        totalOrders:
          type: integer
        totalSpent:
          type: number
        createdAt:
          type: string
...