# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AJN Luxury Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  title: string;
  customerName: string;
  customerPhone: string;
  serviceType?: string;
  date: string;
  /** @nullable */
  time?: string | null;
  status: BookingStatus;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}