# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AJN Luxury Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  name: string;
  fullName: string;
  username: string;
  phone?: string;
  role: UserRole;
  isActive: boolean;
  permissions: string[];
  createdAt?: string;
  updatedAt?: string;
}